+++ src/Insider.Loader/PluginDependencyDescriptor.cs
@@ -1,0 +1,14 @@
+namespace Insider.Loader;
+
+public sealed class PluginDependencyDescriptor
+{
+    internal PluginDependencyDescriptor(string id, bool optional)
+    {
+        Id = id;
+        Optional = optional;
+    }
+
+    public string Id { get; }
+
+    public bool Optional { get; }
+}